+++ lib/src/model/referral_transaction.dart
@@ -5,6 +5,7 @@
 // ignore_for_file: unused_element
 import 'package:built_collection/built_collection.dart';
 import 'package:forestvpn_api/src/model/amount_rate.dart';
+import 'package:forestvpn_api/src/model/error.dart';
 import 'package:built_value/built_value.dart';
 import 'package:built_value/serializer.dart';
 
@@ -16,7 +17,10 @@ part 'referral_transaction.g.dart';
 /// * [id] 
 /// * [currency] 
 /// * [amount] 
+/// * [pending] 
+/// * [pendingExpiresAt] 
 /// * [rates] 
+/// * [errors] 
 /// * [createdAt] 
 @BuiltValue()
 abstract class ReferralTransaction implements Built<ReferralTransaction, ReferralTransactionBuilder> {
@@ -29,8 +33,17 @@ abstract class ReferralTransaction implements Built<ReferralTransaction, Referra
   @BuiltValueField(wireName: r'amount')
   double get amount;
 
+  @BuiltValueField(wireName: r'pending')
+  bool get pending;
+
+  @BuiltValueField(wireName: r'pending_expires_at')
+  DateTime get pendingExpiresAt;
+
   @BuiltValueField(wireName: r'rates')
   BuiltList<AmountRate> get rates;
+
+  @BuiltValueField(wireName: r'errors')
+  BuiltList<Error> get errors;
 
   @BuiltValueField(wireName: r'created_at')
   DateTime get createdAt;
@@ -73,10 +86,25 @@ class _$ReferralTransactionSerializer implements PrimitiveSerializer<ReferralTra
       object.amount,
       specifiedType: const FullType(double),
     );
+    yield r'pending';
+    yield serializers.serialize(
+      object.pending,
+      specifiedType: const FullType(bool),
+    );
+    yield r'pending_expires_at';
+    yield serializers.serialize(
+      object.pendingExpiresAt,
+      specifiedType: const FullType(DateTime),
+    );
     yield r'rates';
     yield serializers.serialize(
       object.rates,
       specifiedType: const FullType(BuiltList, [FullType(AmountRate)]),
+    );
+    yield r'errors';
+    yield serializers.serialize(
+      object.errors,
+      specifiedType: const FullType(BuiltList, [FullType(Error)]),
     );
     yield r'created_at';
     yield serializers.serialize(
@@ -127,12 +155,33 @@ class _$ReferralTransactionSerializer implements PrimitiveSerializer<ReferralTra
           ) as double;
           result.amount = valueDes;
           break;
+        case r'pending':
+          final valueDes = serializers.deserialize(
+            value,
+            specifiedType: const FullType(bool),
+          ) as bool;
+          result.pending = valueDes;
+          break;
+        case r'pending_expires_at':
+          final valueDes = serializers.deserialize(
+            value,
+            specifiedType: const FullType(DateTime),
+          ) as DateTime;
+          result.pendingExpiresAt = valueDes;
+          break;
         case r'rates':
           final valueDes = serializers.deserialize(
             value,
             specifiedType: const FullType(BuiltList, [FullType(AmountRate)]),
           ) as BuiltList<AmountRate>;
           result.rates.replace(valueDes);
+          break;
+        case r'errors':
+          final valueDes = serializers.deserialize(
+            value,
+            specifiedType: const FullType(BuiltList, [FullType(Error)]),
+          ) as BuiltList<Error>;
+          result.errors.replace(valueDes);
           break;
         case r'created_at':
           final valueDes = serializers.deserialize(