+++ test/referral_program_terms_test.dart
@@ -1,0 +1,35 @@
+import 'package:test/test.dart';
+import 'package:forestvpn_api/forestvpn_api.dart';
+
+// tests for ReferralProgramTerms
+void main() {
+  final instance = ReferralProgramTermsBuilder();
+  // TODO add properties to the builder and call build()
+
+  group(ReferralProgramTerms, () {
+    // Initial plan purchase percentage.
+    // double initialPct
+    test('to test the property `initialPct`', () async {
+      // TODO
+    });
+
+    // Subsequent plan purchase percentage.
+    // double subsequentPct
+    test('to test the property `subsequentPct`', () async {
+      // TODO
+    });
+
+    // Mining coins percentage.
+    // double miningPct
+    test('to test the property `miningPct`', () async {
+      // TODO
+    });
+
+    // The duration within which rewards can redeem if the referral program criteria have not been met.
+    // String pendingDur
+    test('to test the property `pendingDur`', () async {
+      // TODO
+    });
+
+  });
+}